# specs/004-crecust-customer-create-modernization/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CRECUST Modernized Customer Create API
  description: |
    REST API for modernized CRECUST customer creation.
    Contract is copybook-constrained and preserves legacy status semantics.
  version: 1.0.0
  contact:
    name: CRECUST Modernization Team

servers:
  - url: http://localhost:8080/v1
    description: Local development environment

tags:
  - name: Customer Create
    description: Customer onboarding and creation

paths:
  /customers:
    post:
      operationId: createCustomer
      summary: Create a new customer
      description: |
        Creates a customer using copybook-backed request fields.
        Customer number is generated by the system from sortcode-scoped control state.
      tags:
        - Customer Create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created successfully
          headers:
            X-Correlation-ID:
              description: Correlation ID for request traceability
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCustomerResponse'
        '400':
          description: Invalid payload or format-level validation failure
          headers:
            X-Correlation-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business-rule validation failure mapped from legacy fail-code paths
          headers:
            X-Correlation-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected internal processing failure
          headers:
            X-Correlation-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Repository/control-state unavailable
          headers:
            X-Correlation-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    DateOfBirth:
      type: object
      required:
        - day
        - month
        - year
      properties:
        day:
          type: integer
          minimum: 1
          maximum: 31
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 1601
          maximum: 9999

    CreatedDate:
      type: object
      required:
        - day
        - month
        - year
      properties:
        day:
          type: integer
          minimum: 1
          maximum: 31
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 1601
          maximum: 9999

    Address:
      type: object
      required:
        - line1
        - city
        - postcode
        - country
      properties:
        line1:
          type: string
          maxLength: 50
        line2:
          type: string
          maxLength: 50
        city:
          type: string
          maxLength: 50
        postcode:
          type: string
          maxLength: 10
        country:
          type: string
          maxLength: 50

    CreateCustomerRequest:
      type: object
      required:
        - title
        - firstName
        - lastName
        - dateOfBirth
        - createdDate
        - phone
        - address
        - status
      properties:
        title:
          type: string
          maxLength: 10
          description: Allowed values are validated by business rules.
          example: Mr
        firstName:
          type: string
          maxLength: 50
          example: John
        lastName:
          type: string
          maxLength: 50
          example: Smith
        dateOfBirth:
          $ref: '#/components/schemas/DateOfBirth'
        createdDate:
          $ref: '#/components/schemas/CreatedDate'
        phone:
          type: string
          maxLength: 20
          example: '4165550101'
        address:
          $ref: '#/components/schemas/Address'
        status:
          type: string
          maxLength: 10

    LegacyStatus:
      type: object
      required:
        - commSuccess
        - commFailCode
      properties:
        commSuccess:
          type: string
          minLength: 1
          maxLength: 1
          example: Y
        commFailCode:
          type: string
          minLength: 0
          maxLength: 1
          example: ' '

    CreateCustomerResponse:
      type: object
      required:
        - eyecatcher
        - sortCode
        - customerNumber
        - title
        - firstName
        - lastName
        - dateOfBirth
        - phone
        - addressLine1
        - addressLine2
        - city
        - postcode
        - country
        - status
        - createdDate
        - creditScore
        - creditScoreReviewDate
        - legacyStatus
      properties:
        eyecatcher:
          type: string
          minLength: 4
          maxLength: 4
          example: CUST
        sortCode:
          type: string
          pattern: '^\d{6}$'
          minLength: 6
          maxLength: 6
          example: '123456'
        customerNumber:
          type: string
          pattern: '^\d{10}$'
          minLength: 10
          maxLength: 10
          example: '0000000006'
        title:
          type: string
          maxLength: 10
        firstName:
          type: string
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        dateOfBirth:
          type: string
          format: date
          example: '1988-03-22'
        phone:
          type: string
          maxLength: 20
        addressLine1:
          type: string
          maxLength: 50
        addressLine2:
          type: string
          maxLength: 50
        city:
          type: string
          maxLength: 50
        postcode:
          type: string
          maxLength: 10
        country:
          type: string
          maxLength: 50
        status:
          type: string
          maxLength: 10
        createdDate:
          type: string
          format: date
          example: '2026-07-22'
        creditScore:
          type: integer
          minimum: 0
          maximum: 999
          example: 742
        creditScoreReviewDate:
          type: string
          format: date
          example: '2026-08-02'
        legacyStatus:
          $ref: '#/components/schemas/LegacyStatus'

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - correlationId
            - timestamp
          properties:
            code:
              type: string
              pattern: '^ERR-\d{3}$'
              example: ERR-101
            message:
              type: string
              example: Invalid title value.
            legacyFailCode:
              type: string
              maxLength: 1
              example: T
            correlationId:
              type: string
              example: 550e8400-e29b-41d4-a716-446655440000
            timestamp:
              type: string
              format: date-time
              example: '2026-07-22T19:00:00Z'